cheri: Fix compiler errors in drivers/char
Fix more compiler errors in drivers/char.

+++ drivers/char/xillybus/xillyusb.c
@@ -519,7 +519,7 @@ static struct xillyusb_endpoint
 
 	for (i = 0; i < bufnum; i++) {
 		struct xillybuffer *xb;
-		unsigned long addr;
+		uintptr_t addr;
 
 		xb = kzalloc(sizeof(*xb), GFP_KERNEL);
